add authentiaction to all REST API

+++ controllers/feed.js
@@ -2,6 +2,7 @@ const { validationResult } = require("express-validator");
 const Post = require("../models/post");
 const fs = require("fs");
 const path = require("path");
+const User = require("../models/user");
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -46,31 +47,36 @@ exports.createPost = (req, res) => {
         throw error;
     }
 
+    const userId = req.userId;
+
+    let creator;
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
         imageUrl: req.file.path,
-        creator: {
-            name: "Haryo"
-        },
+        creator: userId,
     });
 
     post.save()
         .then(result => {
+            return User.findById(userId);
+        })
+        .then(user => {
+            if (!user) {
+                const error = new Error('User not found');
+                error.statusCode = 404;
+                throw error;
+            }
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({
-                message: "Post created successfully",
-                post: {
-                    id: result._id,
-                    title: result.title,
-                    content: result.content,
-                    createdAt: result.createdAt,
-                    imageUrl: result.imageUrl,
-                    creator: {
-                        name: "Haryo",
-                        avatar: "images/IGM.jpg",
-                    },
-                },
-            });
+                message: "Post Created",
+                post: post,
+                creator: { _id: creator._id, name: creator.name },
+            })
         })
         .catch(err => {
             if (!err.statusCode) {
@@ -128,6 +134,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
             if (post.imageUrl !== imageUrl) {
                 clearImage(post.imageUrl);
             }
@@ -156,10 +167,23 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString() !== req.userId) {
+                const error = new Error('Not authorized');
+                error.statusCode = 403;
+                throw error;
+            }
+
+            clearImage(post.imageUrl);
             return Post.deleteOne({ _id: postId });
         })
         .then((post) => {
-            clearImage(post.imageUrl);
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then(result => {
             res.status(200).json({ message: "Post deleted successfully" });
         })
         .catch(err => {
